sw_redlinehide_4b: fix InsertCnt_() iteration of hidden tables
The iteration was stopped as soon as the first start-node of a table
cell was reached, because the SwTableNode case didn't skip over the
table section in the nodes-array.  As seen in ooo67621-1.odt.

It looks like the same problem can't happen with sections or flys
because only tables have nested SwStartNodes and the Fly's nodes
can hardly all be in a delete redline.

Change-Id: Ib21a758c588d5dc8193ba171353f872de71bd65a