tdf#120676: don't send "processing done" when refused to process
After commit 95ab04184ec3fcd9780e7e62babd083d484793d4, condition
RequestHandler::cProcessed is set regardless if the request was
processed in RequestHandler::ExecuteCmdLineRequests. This allows
PipeIpcThread::execute to avoid infinite wait for the condition
object that is never signaled. But this makes the latter function
to send PROCESSING_DONE to the calling process. So the secondary
process gets no hint that the request actually failed.

This change adds a boolean field to signal if the request really
was processed. In the case of refused processing (i.e., when
pGlobal.is() is false, or when !pGlobal->AreRequestsEnabled() is
true, like when shutting down), the secondary process would get
empty response to its request in PipeIpcThread::enable, instead
of expected PROCESSING_DONE, and thus the process will eventually
exit with non-0 error code, which would allow to detect this in
scripts.

Change-Id: Id91ea28025e3f3ab60c7049f02ed9fc226d2a654
Reviewed-on: https://gerrit.libreoffice.org/65610
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>