oox smartart, org chart: handle multiple paragraphs on data node
This problem was similar to the one fixed in
2ee590f9eae5992af5137b6376b1ff820b4f6333 (oox smartart, accent process:
handle multiple runs from a data point, 2018-11-21), but this there we
handled multiple runs and this handles multiple paragraphs.

It seems some smartart types allow multiple paragraphs in a diagram
node, others only allow multiple runs. Org chart is in the former
category.

Change-Id: I281f01fdfa809d0a232d5da7fdaa23de7adcd627
Reviewed-on: https://gerrit.libreoffice.org/66066
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins