Related: tdf#122509 prevent access of not existing base format
Call to pFormat->HasPositiveBracketPlaceholder() with nullptr as
obtained for the not existing nIndex was the immediate cause of
the crash.

Also obtain the type from the format if found instead of the
separate GetType() call that again would need to find the format
first. This clarifies that the other places accessing pFormat
don't need additional checks because eType isn't UNDEFINED there.

Change-Id: I5d78ad73eeb4c3820c45d6a91b0e729ed22951a4
Reviewed-on: https://gerrit.libreoffice.org/66476
Reviewed-by: Eike Rathke <erack@redhat.com>
Tested-by: Jenkins