Add a dummy WordBasic.ToolsOptionsView
Does nothing. Needed for customer application to proceed. Once we are
further along in getting it to work, we can investigate what the
parameters passed to this ToolsOptionsView method actually are. (This
WordBasic thing is something that has been deprecated since last
century, I suspect, so no wonder it is hard to find authoritative
documentation on it.)

Change-Id: I62a6d6d9abb9364afca110570fa341a2375a77a6