tdf#90097 rtfimport: rewrite unit test so it run reliably
Already broken for windows and mac, the numbers changed
for me on Ubuntu when switching from gcc5 to gcc7 or gcc8.
How it still  works on the tinderboxes etc I don't know.
Anyway, by removing the obviously irrelevant actual numbers,
the test still covers the original intent by ensuring that
the top point is farther to the right than the bottom point.

Change-Id: Ieeb7cd854847d47a11dc305db82371e2b8f85049
Reviewed-on: https://gerrit.libreoffice.org/66724
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>