tdf#122878: enable wrap for flys in footer
This patch removes the check for a footer node,
intoduced by
23f698ecee033612ac3a9f5cfd7674b08bb3ccd1
preserving the behaviour for the connected bug reports
i13832 and i24135.

Without this check, the wraping becomes enabled for footer
objects, too.

With this enhencement, the commits
95baa7f4fdaa5b857efcb82f32854b06bd37ddf8
d519fb1b4bea10af513cc56fe1b0059a9824074a
are directly affected and therefore the unit test is edited.

Change-Id: I093add9e251ac97859a66cb8b8563010ef734c2d
Reviewed-on: https://gerrit.libreoffice.org/67069
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>