Assert missing SalGraphics font on GetTextLayout
Trying to layout text without a font can't work.
Without a font, currently the text is simply not painted, but this
is still a programming error, so warn the developer.

And we can get rid of all the intermediate OutputDevice font and
gpahics checks, as everything now uses Harfbuzz and there won't
be any layout which can't use the cached runs, so just generate
them directly.

Change-Id: Iae1fab0afeefec1cb1ca82c4e5d8dcbe9919859d
Reviewed-on: https://gerrit.libreoffice.org/66818
Tested-by: Jenkins
Reviewed-by: Jan-Marek Glogowski <glogow@fbihome.de>