tdf#122259 resolve image link as both - png and svg
This is needed to prevent mismatches when we have a image defined
as png and we have a svg available. As we currently support both
png and svg icon themes we need to try both or some resolvable
icons just won't be resolved. Once svg support is in a good shape
that we can get rid of png, we can also possibly get rid of this.

Change-Id: I34e9c24ac1016ef3af46be06330b60aaab730da0
Reviewed-on: https://gerrit.libreoffice.org/66826
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>