Drop kde4 desktop detection
KDE 4 has been unmaintained upstream for various year now and
the kde4 VCL Plugin was removed with
7c95f1f5247cbaeb38c28f919e2795e6884e8a99 in favor of the kde5 one.

Drop desktop detection for this obsolete KDE 4 version as well.
Should new LO versions be run on old systems still running KDE 4, it's
very unlikely that they have the kf5 libraries that are required
to use the kde5 or gtk3_kde5 SAL plugins anyway, so the default
fallback for an unknown desktop environment is quite as good.

Change-Id: If8de489fbca836a80451435edfb00d2a8d501031
Reviewed-on: https://gerrit.libreoffice.org/67053
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>