tdf#123002 Disable off-page positioning for footer/header objects only
We should restrict this to this case since the anchor of objects in the
body may appear on another page when the object is moved around.

In constrast, objects in header and footer should not appear on other pages,
so we still disable the "off-page positioning" for them.

Horizontal off-page positioning should be disabled in any case.

See also
tdf#112443 and tdf#120839

Change-Id: I056c74526f38c302ba49297f9f84ec0e958d2cec
Reviewed-on: https://gerrit.libreoffice.org/67088
Tested-by: Jenkins
Tested-by: Xisco Faulí <xiscofauli@libreoffice.org>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>