framework: allow dispatching a command on the main thread
This is similar to commit ebbc0aa020cdba478fbf52d3f3ae8289e1c37906
(framework: allow loading a component on the main thread, 2018-12-19),
just it allows saving (via .uno:Save) and other commands operating in a
similar environment.

The use-case is that once a document is loaded on the main thread (see
commit message of the above mentioned commit), then saving also has to
happen on the main thread, or OLE objects on Windows may be lost.

Change-Id: I7321659550b556e96085ac20f197a87d5d13f1ed
Reviewed-on: https://gerrit.libreoffice.org/67089
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins