framework: allow storeSelf() on the main thread
This is similar to commit ed07d92e16f678e229f6775a4a4e401500e57ce1
(framework: allow dispatching a command on the main thread, 2019-01-29),
except it addresses the scenario where save is performed with an UNO API
call, not with a dispatched command.

This way both load and save is possible on the main thread, providing a
safe way to interact with documents containing OLE objects on Windows.

Change-Id: I3516c944ece8ed0e26aa13fc9def5857b8344404
Reviewed-on: https://gerrit.libreoffice.org/67109
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins