oox: disable gradient fill grab-bag for PPTX
This was added in commit 7f3762a70b15ea2ecebada02029d994580372f23
(ooxml: preserve gradient shape fill, 2014-01-31), and assumes that the
theme colors can be preserved, as the theme definition is grab-bagged as
well.

But the theme is grab-bagged only for DOCX, not for PPTX, so skip
gradient grab-bag for PPTX, otherwise the gradient would refer to
incorrect colors in the theme.

Change-Id: I98e1c67d4b10e68916f81dd7fc508eb4146d506b
Reviewed-on: https://gerrit.libreoffice.org/67386
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins