tdf#122607 Fix layout invalidation in headless mode
Layout was not invalidated correctly since the visible area is 0
in headless mode.
So just reformat the whole doc when doing the pdf conversion headless.

An attempt to fix this was already made with commit 48d83d3cfeee165149d1bd262684f5daf7ae37e8
but that didn't cover all cases.

Change-Id: I3f620b2f2db2c4a6e5bf279b33e5c93697e4e2d4
Reviewed-on: https://gerrit.libreoffice.org/67417
Tested-by: Jenkins
Reviewed-by: Samuel Mehrbrodt <Samuel.Mehrbrodt@cib.de>